refactor(services): move findFrontmatterEndOffset to a frontmatter-offset leaf (#1417)
feature-definition.ts — the shared leaf helper for the markdown-defined
feature managers — reached back up into skill-manager.ts to import
findFrontmatterEndOffset, a 15-line pure scanner. That forced the leaf to
depend on a 491-line manager module (Obsidian APIs, plugin type surface,
tool-policy types, bundled-skill registry), and four separate test files
had to vi.mock the entire manager just to control one dependency-free
function — two of them while testing modules that never mention skills.

Move the scanner verbatim to a new leaf module, src/utils/
frontmatter-offset.ts, alongside bundled-frontmatter.ts (whose header
already documents that the offset-based vault-file scanner and the
regex-based bundled reader are deliberately separate paths — this moves
one, it does not merge them). skill-manager.ts re-exports the symbol so
existing import paths keep working; feature-definition.ts and the
skill-manager's own internal caller now import the leaf directly.

Also corrects the same reach-back in skill-mention-modal.ts: SkillSummary
is imported from the skill-types leaf that exists for exactly that
purpose, instead of from skill-manager.

Test retargeting: feature-definition.test.ts and hook-manager.test.ts
mock the new leaf (they drive its return value); the defensive mocks in
file-backed-feature-manager.test.ts and scheduled-task-manager.test.ts
are deleted outright — the real function is pure and safe to run.
skill-manager.test.ts's import path is unchanged via the re-export.
